Use dynamic import to load codemirror
